+++ src/main/java/id/naturalsmp/naturalSchool/ui/UIManager.java
@@ -25,7 +25,7 @@ import java.util.logging.Level;
 
 public class UIManager {
 
-    public static final String GUI_VERSION = "1.7.5";
+    public static final String GUI_VERSION = "1.7.6";
 
     private static final MiniMessage MM = MiniMessage.miniMessage();